chore: Move hooks to hook folder
