+++ source/sites/publicodes/Action.js
@@ -3,7 +3,6 @@ import { splitName } from 'Components/publicodesUtils'
 import Simulation from 'Components/Simulation'
 import { Markdown } from 'Components/utils/markdown'
 import { ScrollToTop } from 'Components/utils/Scroll'
-import { useNextQuestions } from 'Components/utils/useNextQuestion'
 import { utils } from 'publicodes'
 import { useContext, useEffect } from 'react'
 import emoji from 'react-easy-emoji'
@@ -13,6 +12,7 @@ import { useParams } from 'react-router'
 import { Link } from 'react-router-dom'
 import { EngineContext } from '../../components/utils/EngineContext'
 import Meta from '../../components/utils/Meta'
+import { useNextQuestions } from '../../hooks/useNextQuestion'
 import { questionConfig } from './questionConfig'
 
 const { decodeRuleName, encodeRuleName } = utils
