+++ source/sites/publicodes/App.tsx
@@ -12,8 +12,8 @@ import AnimatedLoader from '../../AnimatedLoader'
 import Footer from '../../components/Footer'
 import LangSwitcher from '../../components/LangSwitcher'
 import LocalisationMessage from '../../components/localisation/LocalisationMessage'
-import useMediaQuery from '../../components/utils/useMediaQuery'
 import { TrackerContext } from '../../contexts/TrackerContext'
+import useMediaQuery from '../../hooks/useMediaQuery'
 import Provider from '../../Provider'
 import { AppState } from '../../reducers/rootReducer'
 import { WithEngine } from '../../RulesProvider'
