+++ source/sites/publicodes/Landing.tsx
@@ -15,8 +15,8 @@ import { CircleSVG } from '../../components/ProgressCircle'
 import { openmojiURL } from '../../components/SessionBar'
 import { IframeOptionsContext } from '../../components/utils/IframeOptionsProvider'
 import Meta from '../../components/utils/Meta'
-import useMediaQuery from '../../components/utils/useMediaQuery'
 import { TrackerContext } from '../../contexts/TrackerContext'
+import useMediaQuery from '../../hooks/useMediaQuery'
 import LandingExplanations from './LandingExplanations'
 import { useProfileData } from './Profil'
