+++ source/sites/publicodes/Navigation.tsx
@@ -3,7 +3,7 @@ import { useContext } from 'react'
 import LangSwitcher from '../../components/LangSwitcher'
 import Logo from '../../components/Logo'
 import { IframeOptionsContext } from '../../components/utils/IframeOptionsProvider'
-import useMediaQuery from '../../components/utils/useMediaQuery'
+import useMediaQuery from '../../hooks/useMediaQuery'
 import SkipLinks from './SkipLinks'
 
 export default ({ fluidLayout }) => {
