+++ source/sites/publicodes/Tutorial.tsx
@@ -8,8 +8,8 @@ import { Link, useNavigate } from 'react-router-dom'
 import { skipTutorial } from '../../actions/actions'
 import SlidesLayout from '../../components/SlidesLayout'
 import Meta from '../../components/utils/Meta'
-import useKeypress from '../../components/utils/useKeyPress'
 import { TrackerContext } from '../../contexts/TrackerContext'
+import useKeypress from '../../hooks/useKeyPress'
 import { WithEngine } from '../../RulesProvider'
 import Chart from './chart/index.js'
 import HorizontalSwipe from './HorizontalSwipe'
