+++ source/sites/publicodes/conference/ConferenceBar.tsx
@@ -7,7 +7,7 @@ import { Link } from 'react-router-dom'
 import { situationSelector } from 'Selectors/simulationSelectors'
 import * as Y from 'yjs'
 import { minimalCategoryData } from '../../../components/publicodesUtils'
-import { useSimulationProgress } from '../../../components/utils/useNextQuestion'
+import { useSimulationProgress } from '../../../hooks/useNextQuestion'
 import { conferenceElementsAdapter } from './Conference'
 import { GroupModeMenuEntryContent } from './GroupModeSessionVignette'
 import { computeHumanMean } from './GroupStats'
