+++ source/sites/publicodes/conference/Survey.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from 'react-router-dom'
 import { Trans, useTranslation } from 'react-i18next'
 import { conferenceImg } from '../../../components/SessionBar'
 import Meta from '../../../components/utils/Meta'
-import { usePersistingState } from '../../../components/utils/persistState'
+import { usePersistingState } from '../../../hooks/usePersistState'
 import Navigation from '../Navigation'
 import { useProfileData } from '../Profil'
 import { ConferenceTitle } from './Conference'
