+++ source/sites/publicodes/conference/SurveyBar.tsx
@@ -4,14 +4,14 @@ import {
 	splitName,
 } from 'Components/publicodesUtils'
 import { useEngine } from 'Components/utils/EngineContext'
-import { usePersistingState } from 'Components/utils/persistState'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
 import { situationSelector } from 'Selectors/simulationSelectors'
 import { v4 as uuidv4 } from 'uuid'
 import { minimalCategoryData } from '../../../components/publicodesUtils'
-import { useSimulationProgress } from '../../../components/utils/useNextQuestion'
+import { useSimulationProgress } from '../../../hooks/useNextQuestion'
+import { usePersistingState } from '../../../hooks/usePersistState'
 import { GroupModeMenuEntryContent } from './GroupModeSessionVignette'
 import { computeHumanMean } from './GroupStats'
 import { surveyElementsAdapter } from './Survey'
