+++ source/sites/publicodes/fin/index.tsx
@@ -6,8 +6,8 @@ import { useDispatch, useSelector } from 'react-redux'
 import { Link, useSearchParams } from 'react-router-dom'
 import { answeredQuestionsSelector } from 'Selectors/simulationSelectors'
 import SlidesLayout from '../../../components/SlidesLayout'
-import { useNextQuestions } from '../../../components/utils/useNextQuestion'
 import { TrackerContext } from '../../../contexts/TrackerContext'
+import { useNextQuestions } from '../../../hooks/useNextQuestion'
 import { arrayLoopIteration } from '../../../utils'
 import EnqueteBannerContent from '../enquête/BannerContent'
 import { enquêteSelector } from '../enquête/enquêteSelector'
